# Api/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Asientos/resumen": {
      "get": {
        "tags": [
          "Asientos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "hasta",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "debId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "creId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AsientoResumenDto"
              }
            }
          }
        }
      }
    },
    "/api/Asientos/detalle": {
      "get": {
        "tags": [
          "Asientos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "anio",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "mes",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "debId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "creId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AsientoDetalleDto"
            }
          }
        }
      }
    },
...
    "/api/Dashboard/cheques-trend": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "months",
            "type": "integer",
            "format": "int32",
            "default": 6
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ChequesTrendPoint"
              }
            }
          }
        }
      }
    },
    "/api/Dashboard/top-proveedores": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "months",
            "type": "integer",
            "format": "int32",
            "default": 6
          },
          {
            "in": "query",
            "name": "take",
            "type": "integer",
            "format": "int32",
            "default": 5
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TopProveedorDto"
              }
            }
          }
        }
      }
    },
    "/api/Proveedores": {
      "get": {
        "tags": [
          "Proveedores"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Proveedores"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Proveedor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Proveedores/{id}": {
      "get": {
        "tags": [
          "Proveedores"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Proveedores"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Proveedor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Proveedores"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AsientoDetalleChequeDto": {
      "type": "object",
      "properties": {
        "numeroSolicitud": {
          "format": "int32",
          "type": "integer"
        },
        "numeroCheque": {
          "type": "string"
        },
        "proveedorId": {
          "format": "int32",
          "type": "integer"
        },
        "proveedorNombre": {
          "type": "string"
        },
        "proveedorRnc": {
          "type": "string"
        },
        "monto": {
          "format": "double",
          "type": "number"
        },
        "fechaRegistro": {
          "format": "date-time",
          "type": "string"
        },
        "cuentaContableProveedor": {
          "type": "string"
        },
        "cuentaContableBanco": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "AsientoDetalleDto": {
      "type": "object",
      "properties": {
        "anio": {
          "format": "int32",
          "type": "integer"
        },
        "mes": {
          "format": "int32",
          "type": "integer"
        },
        "debId": {
          "format": "int32",
          "type": "integer"
        },
        "creId": {
          "format": "int32",
          "type": "integer"
        },
        "montoTotal": {
          "format": "double",
          "type": "number"
        },
        "cantidadCheques": {
          "format": "int32",
          "type": "integer"
        },
        "proveedoresUnicos": {
          "format": "int32",
          "type": "integer"
        },
        "cheques": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AsientoDetalleChequeDto"
          }
        }
      },
      "additionalProperties": false
    },
    "AsientoResumenDto": {
      "type": "object",
      "properties": {
        "anio": {
          "format": "int32",
          "type": "integer"
        },
        "mes": {
          "format": "int32",
          "type": "integer"
        },
        "debId": {
          "format": "int32",
          "type": "integer"
        },
        "creId": {
          "format": "int32",
          "type": "integer"
        },
        "montoTotal": {
          "format": "double",
          "type": "number"
        },
        "cantidadCheques": {
          "format": "int32",
          "type": "integer"
        },
        "proveedoresUnicos": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "ChequesTrendPoint": {
      "type": "object",
      "properties": {
        "anio": {
          "format": "int32",
          "type": "integer"
        },
        "mes": {
          "format": "int32",
          "type": "integer"
        },
        "monto": {
          "format": "double",
          "type": "number"
        },
        "cantidad": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "Proveedor": {
      "type": "object",
      "properties": {
        "identificador": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "tipoPersona": {
          "type": "string"
        },
        "cedulaRnc": {
          "type": "string"
        },
        "balance": {
          "format": "double",
          "type": "number"
        },
        "cuentaContable": {
          "type": "string"
        },
        "estado": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
    "TopProveedorDto": {
      "type": "object",
      "properties": {
        "proveedorId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "montoTotal": {
          "format": "double",
          "type": "number"
        },
        "cheques": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    }
  }
}